Tycho ITs dev mode support, i.e. running from with m2e workspace
* split WorkspaceDependencyResolver in two classes,
  DevWorkspaceResolver and WorkspaceP2RuntimeLocator.

* added support for osgi.dev to sisu-equinox-launching

* changed TestMojo to resolve junit runtime from dev workspace

* changed AbstractTychoIntegrationTest to run Verifier without
  forking maven in separate jvm and pass required parameters to
  enable m2e and tycho workspace dependency resolution.

+++ sisu-equinox/sisu-equinox-embedder/src/main/java/org/eclipse/sisu/equinox/embedder/internal/DefaultEquinoxEmbedder.java
@@ -107,6 +107,10 @@ public class DefaultEquinoxEmbedder extends AbstractLogEnabled implements Equino
                 }
                 bundleLocations.add(location);
             }
+
+            public void addBundleStartLevel(String id, int level, boolean autostart) {
+                // TODO do we need to autostart?
+            }
         });
 
         if (installationLocations.isEmpty() && !platformProperties.containsKey("osgi.install.area")) {
@@ -308,10 +312,12 @@ public class DefaultEquinoxEmbedder extends AbstractLogEnabled implements Equino
                 getLogger().error("Exception while shutting down equinox", e);
             }
             tempSecureStorage.delete();
-            try {
-                FileUtils.deleteDirectory(tempConfigDir);
-            } catch (IOException e) {
-                getLogger().warn("Exception while deleting temp folder " + tempConfigDir, e);
+            if (tempConfigDir != null) {
+                try {
+                    FileUtils.deleteDirectory(tempConfigDir);
+                } catch (IOException e) {
+                    getLogger().warn("Exception while deleting temp folder " + tempConfigDir, e);
+                }
             }
             frameworkContext = null;
         }
